chore: Fix release build
