+++ Prose/ProseLib/Sources/PreviewAssets/ImageAsset+URL.swift
@@ -18,16 +18,3 @@ public extension ImageAsset {
         return components.url
     }
 }
-
-// TODO: [Rémi Bardon] Here is probably not the right place to put this function, we should move it somewhere else.
-public func assetData(from url: URL) -> (String, Bundle?)? {
-    guard url.scheme == "asset" else { return nil }
-    guard var components = URLComponents(url: url, resolvingAgainstBaseURL: false) else { return nil }
-
-    let query = components.queryItems ?? []
-    components.queryItems = nil
-
-    guard let imageName = query.first(where: { $0.name == "imageName" })?.value else { return nil }
-
-    return (imageName, Bundle(path: components.path))
-}
